// frontend/src/types/AdminAccountDto.ts
/**
 * Shoppy
 * Common spec
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { AccountRole } from './AccountRole';
import type { Status } from './Status';
// import { HttpFile } from '../http/http';

export class AdminAccountDto {
    'id'?: number;
    'login'?: string;
    'password'?: string;
    'salted'?: boolean;
    'status'?: Status;
    'accountRoles'?: Array<AccountRole>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "login",
            "baseName": "login",
            "type": "string",
            "format": ""
        },
        {
            "name": "password",
            "baseName": "password",
            "type": "string",
            "format": ""
        },
        {
            "name": "salted",
            "baseName": "salted",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "Status",
            "format": ""
        },
        {
            "name": "accountRoles",
            "baseName": "accountRoles",
            "type": "Array<AccountRole>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AdminAccountDto.attributeTypeMap;
    }

    public constructor() {
    }
}