// frontend/src/types/AdminOrderDto.ts
/**
 * Shoppy
 * Common spec
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { Status } from './Status';
// import { HttpFile } from '../http/http';

export class AdminOrderDto {
    'id'?: number;
    'info'?: string;
    'status'?: Status;
    'guid'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "info",
            "baseName": "info",
            "type": "string",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "Status",
            "format": ""
        },
        {
            "name": "guid",
            "baseName": "guid",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AdminOrderDto.attributeTypeMap;
    }

    public constructor() {
    }
}