// frontend/src/types/Status.ts
/**
 * Shoppy
 * Common spec
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// import { HttpFile } from '../http/http';

export type Status = "ADDED" | "ACTIVE" | "REMOVED" | "DISABLED" ;